My Eclipse Workspace for Intel Edison development
This is my WIP development for the Intel Edison (and other projects) for
the Intel Edison.

First Drop.

Warning this is WIP, did not include all support libraries that I copy
from Edison.  Will be adding readme and more about it soon!

Feel free to play, but no promises about it being good for anything!