TFT Console - Textbox can scroll by touch...
Put in first version of being able to touch in the scrollable text area
of the run screen, and drag finger up or down to scroll text.  Will add
more scrolling later by maybe timed.  That is if you scroll up and hold,
it will continue to scroll.  Also want to add some visual clue which
areas are scrollable.

Also added <DBG> and <SRV> commands to run screen, first toggles debug
on or off from Phoenix APP, second sends command to show positions of
the servos, to debug if servos are working or not.